refactor(tidal): extract download subsystem into tidal_downloads module
The download-integrity fixes pushed app.py past its frozen file-size
baseline (2354 > 2209). Move the download path — path building and
containment, FLAC finalization, track and album download loops, and the
shared DASH manifest helpers — into tidal_downloads.py with a strict
app.py -> tidal_downloads dependency direction, and tighten the baseline
to the new 1907-line count. Pure move; no behavior change.
