+++ package.json
@@ -20,7 +20,7 @@
         "verify:frontend": "npm --prefix frontend run lint && npm --prefix frontend run build && npm --prefix frontend run test:coverage && npm --prefix frontend run test:component && npm --prefix frontend run typecheck",
         "verify:helm": "./scripts/helm-chart-render-check.sh",
         "verify:python": "pytest services/tidal-streamer/tests -q && pytest services/ytmusic-streamer/tests -q && pytest services/audio-analyzer/tests -q && pytest services/vibe-provider-dclap/tests -q && pytest tests/ -q",
-        "verify:python-quality": "ruff check services/ && ruff format --check services/ && mypy && MYPYPATH=services mypy services/tidal-streamer/app.py services/tidal-streamer/tests && MYPYPATH=services mypy services/ytmusic-streamer/app.py services/ytmusic-streamer/yt_download.py services/ytmusic-streamer/ytmusic_album_downloads.py services/ytmusic-streamer/ytmusic_auth.py services/ytmusic-streamer/ytmusic_browse.py services/ytmusic-streamer/ytmusic_client.py services/ytmusic-streamer/ytmusic_downloads.py services/ytmusic-streamer/ytmusic_library.py services/ytmusic-streamer/ytmusic_lifecycle.py services/ytmusic-streamer/ytmusic_models.py services/ytmusic-streamer/ytmusic_runtime.py services/ytmusic-streamer/ytmusic_search.py services/ytmusic-streamer/ytmusic_stream.py services/ytmusic-streamer/tests && MYPYPATH=services mypy services/audio-analyzer/analyzer.py services/audio-analyzer/loudness.py services/audio-analyzer/model_paths.py && MYPYPATH=services mypy services/vibe-provider-dclap/__main__.py services/vibe-provider-dclap/http_server.py services/vibe-provider-dclap/inference.py services/vibe-provider-dclap/model_provider.py services/vibe-provider-dclap/music_path.py services/vibe-provider-dclap/preprocessing.py services/vibe-provider-dclap/settings.py",
+        "verify:python-quality": "ruff check services/ && ruff format --check services/ && mypy && MYPYPATH=services mypy services/tidal-streamer/app.py services/tidal-streamer/tidal_downloads.py services/tidal-streamer/tests && MYPYPATH=services mypy services/ytmusic-streamer/app.py services/ytmusic-streamer/yt_download.py services/ytmusic-streamer/ytmusic_album_downloads.py services/ytmusic-streamer/ytmusic_auth.py services/ytmusic-streamer/ytmusic_browse.py services/ytmusic-streamer/ytmusic_client.py services/ytmusic-streamer/ytmusic_downloads.py services/ytmusic-streamer/ytmusic_library.py services/ytmusic-streamer/ytmusic_lifecycle.py services/ytmusic-streamer/ytmusic_models.py services/ytmusic-streamer/ytmusic_runtime.py services/ytmusic-streamer/ytmusic_search.py services/ytmusic-streamer/ytmusic_stream.py services/ytmusic-streamer/tests && MYPYPATH=services mypy services/audio-analyzer/analyzer.py services/audio-analyzer/loudness.py services/audio-analyzer/model_paths.py && MYPYPATH=services mypy services/vibe-provider-dclap/__main__.py services/vibe-provider-dclap/http_server.py services/vibe-provider-dclap/inference.py services/vibe-provider-dclap/model_provider.py services/vibe-provider-dclap/music_path.py services/vibe-provider-dclap/preprocessing.py services/vibe-provider-dclap/settings.py",
         "verify:scale": "npm run verify:scale:backend && npm run verify:scale:python",
         "verify:scale:backend": "npm --prefix backend run test:scale",
         "verify:scale:python": "pytest services/audio-analyzer/tests-scale -q && pytest services/vibe-provider-dclap/tests-scale -q",
