+++ scripts/ci/check-file-size.mjs
@@ -75,7 +75,7 @@ const BASELINE = Object.freeze({
     "frontend/lib/audio-controls-context.tsx": 2311,
     "frontend/lib/listen-together-context.tsx": 2073,
     "services/audio-analyzer/analyzer.py": 2266,
-    "services/tidal-streamer/app.py": 2209,
+    "services/tidal-streamer/app.py": 1907,
 });
 
 function packageScanRoots(repoRoot) {
