+++ services/tidal-streamer/tests/test_download_path_safety.py
@@ -44,30 +44,30 @@ def _configure_download(
 def test_download_rejects_absolute_rendered_template(
     tmp_path: Path, monkeypatch: pytest.MonkeyPatch
 ) -> None:
-    import app
+    import tidal_downloads
 
     destination = tmp_path / "music"
     outside_stem = tmp_path / "outside" / "track"
-    _configure_download(monkeypatch, app, outside_stem.as_posix())
+    _configure_download(monkeypatch, tidal_downloads, outside_stem.as_posix())
 
     with pytest.raises(ValueError, match="output template path"):
-        app._download_track_sync(_FakeDownloadApi(), 1, "HIGH", "ignored", destination)
+        tidal_downloads._download_track_sync(_FakeDownloadApi(), 1, "HIGH", "ignored", destination)
 
     assert not outside_stem.with_suffix(".m4a").exists()
 
 
 def test_download_rejects_symlink_escape(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    import app
+    import tidal_downloads
 
     destination = tmp_path / "music"
     outside = tmp_path / "outside"
     destination.mkdir()
     outside.mkdir()
     (destination / "linked").symlink_to(outside, target_is_directory=True)
-    _configure_download(monkeypatch, app, "linked/track")
+    _configure_download(monkeypatch, tidal_downloads, "linked/track")
 
     with pytest.raises(ValueError, match="outside MUSIC_PATH"):
-        app._download_track_sync(_FakeDownloadApi(), 1, "HIGH", "ignored", destination)
+        tidal_downloads._download_track_sync(_FakeDownloadApi(), 1, "HIGH", "ignored", destination)
 
     assert not (outside / "track.m4a").exists()
 
@@ -75,7 +75,7 @@ def test_download_rejects_symlink_escape(tmp_path: Path, monkeypatch: pytest.Mon
 def test_download_rejects_temporary_file_symlink_escape(
     tmp_path: Path, monkeypatch: pytest.MonkeyPatch
 ) -> None:
-    import app
+    import tidal_downloads
 
     destination = tmp_path / "music"
     target_parent = destination / "Artist" / "Album"
@@ -83,10 +83,10 @@ def test_download_rejects_temporary_file_symlink_escape(
     target_parent.mkdir(parents=True)
     outside_file.write_bytes(b"unchanged")
     (target_parent / "01. Track.m4a.tmp").symlink_to(outside_file)
-    _configure_download(monkeypatch, app, "Artist/Album/01. Track", b"replacement")
+    _configure_download(monkeypatch, tidal_downloads, "Artist/Album/01. Track", b"replacement")
 
     with pytest.raises(ValueError, match="outside MUSIC_PATH"):
-        app._download_track_sync(_FakeDownloadApi(), 1, "HIGH", "ignored", destination)
+        tidal_downloads._download_track_sync(_FakeDownloadApi(), 1, "HIGH", "ignored", destination)
 
     assert outside_file.read_bytes() == b"unchanged"
 
@@ -94,12 +94,14 @@ def test_download_rejects_temporary_file_symlink_escape(
 def test_download_writes_valid_rendered_template_beneath_destination(
     tmp_path: Path, monkeypatch: pytest.MonkeyPatch
 ) -> None:
-    import app
+    import tidal_downloads
 
     destination = tmp_path / "music"
-    _configure_download(monkeypatch, app, "Artist/Album/01. Track")
+    _configure_download(monkeypatch, tidal_downloads, "Artist/Album/01. Track")
 
-    result = app._download_track_sync(_FakeDownloadApi(), 1, "HIGH", "ignored", destination)
+    result = tidal_downloads._download_track_sync(
+        _FakeDownloadApi(), 1, "HIGH", "ignored", destination
+    )
 
     expected = destination / "Artist" / "Album" / "01. Track.m4a"
     assert expected.read_bytes() == b"audio"
@@ -114,11 +116,12 @@ async def test_download_route_sanitizes_component_that_becomes_empty(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     import app
+    import tidal_downloads
 
     destination = tmp_path / "music"
     outside_stem = tmp_path / "outside" / "escaped"
     rendered_path = f"..{outside_stem.as_posix()}"
-    _configure_download(monkeypatch, app, rendered_path)
+    _configure_download(monkeypatch, tidal_downloads, rendered_path)
     monkeypatch.setattr(app, "MUSIC_PATH", destination)
     monkeypatch.setattr(app, "_build_api", lambda *_args: _FakeDownloadApi())
